==> Perfil_Laranja/02_Join_dos_Dados.ipynb <==
# Importando dependencias

import warnings
warnings.filterwarnings("ignore")

# Utils

BASE_GS = "gs://oculto/loss_prevention/perfil_laranja/data/"

# Pré-Tratamento dos Dados

arquivos = [
"bases de dados foram ocultadas para preservar a privacidade dos dados",
]

df_final = spark.read.parquet(arquivos[0])

for arquivo in arquivos[1:]:
    df_temp = spark.read.parquet(arquivo)
    df_final = df_final.join(df_temp, on="cpf", how="left")

df_final = df_final.fillna(0)

df_final.write.mode("overwrite").parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_unificado.parquet")


df_final.limit(5).toPandas()

df_final.count()
