==> Perfil_Laranja/04.5_pre_processing_com_cpfs.ipynb <==
# Importando dependencias

import pyspark.sql.functions as sf
from pyspark.sql.window import Window
from pyspark.sql.types import IntegerType
from feature_store import FeatureStore, Catalog
from pyspark.sql import DataFrame
import time
from pyspark.sql.functions import col, when, lit

# Pré-Tratamento dos Dados

dataset = spark.read.parquet('gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_unificado.parquet')

dataset.count()

## Apagar colunas irrelevantes

dt_treino = dataset.drop("id","insertDate","nome","idade","dataIndicioFalecimento","situacao","quantidadeRelacionamentos","processingDateB3","referencia","referenciaAnoMes")

dt_treino = dt_treino.drop("flagSituacaoRegular")

## Tratar colunas categoricas

anos = {
"SEM INFORMACAO": "-1",
"SEM RECORRENCIA": "0",
"ATE 30 DIAS": "1",
"DE 30 A 90 DIAS": "2",
"DE 90 A 180 DIAS": "3",
"DE 180 DIAS A 1 ANO": "4",
"DE 1 A 2 ANOS": "5",
"DE 2 A 3 ANOS": "6",
"DE 3 A 5 ANOS": "7",
"DE 5 A 10 ANOS": "8",
"MAIS QUE 10 ANOS": "9",
}

letras = {
    "SEM INFORMACAO": "-1",
    "A": "1",
    "B": "2",
    "C": "3",
    "D": "4",
    "E": "5",
    "F": "6",
    "G": "7",
    "H": "8",
    "I": "9",
    "J": "10",
    "K": "11",
    "L": "12",
    "M": "13",
    "N": "14",
    "O": "15",
    "P": "16",
    "Q": "17",
    "R": "18",
    "S": "19",
    "T": "20",
    "U": "21",
    "V": "22",
    "W": "23",
    "X": "24",
    "Y": "25",
    "Z": "26"
}

classe1 = {
    "SEM INFORMACAO": "-1",
    "BAIXO": "1",
    "MEDIO": "2",
    "ALTO": "3"
}

classe2 = {
    "SEM INFORMACAO": "-1",
    "BAIXA": "1",
    "MEDIA": "2",
    "ALTA": "3"
}

colunas_string = [f.name for f in dt_treino.schema.fields if (f.dataType.simpleString() == 'string')&(f.dataType.simpleString() != 'boolean')]

dt_treino = dt_treino.fillna("SEM INFORMACAO", subset=colunas_string)

mapeamentos = {
    "recorrenciaFinanciamentoVeicular": anos,
    "ativosComCota": letras,
    "ativosSemCota": letras,
    "bancarizacao": letras,
    "bancarizacaoBest": classe1,
    "bancarizacaoFaixa": classe2,
    "diversificacao": letras,
    "diversificacaoBest": classe1,
    "diversificacaoFaixa": classe2,
    "patrimonio": letras
}

# Aplica os replaces de forma dinâmica
for coluna, mapa in mapeamentos.items():
    dt_treino = dt_treino.replace(mapa, subset=[coluna])

    dt_treino = dt_treino.withColumn(coluna, col(coluna).cast("int"))



dt_treino = (
    dt_treino
    .withColumn("historicoAtrasoFinanciamentoVeicularSEM_INFORMACAO", (col("historicoAtrasoFinanciamentoVeicular") == "SEM INFORMACAO").cast("double"))
    .withColumn("historicoAtrasoFinanciamentoVeicularATRASA", (col("historicoAtrasoFinanciamentoVeicular") == "ATRASA").cast("double"))
    .withColumn("historicoAtrasoFinanciamentoVeicularNAO_ATRASA", (col("historicoAtrasoFinanciamentoVeicular") == "NAO ATRASA").cast("double"))
)

dt_treino = (
    dt_treino
    .withColumn("evolucaoPatrimonioSEM_INFORMACAO", (col("evolucaoPatrimonio") == "SEM INFORMACAO").cast("double"))
    .withColumn("evolucaoPatrimonioAUMENTOU", (col("evolucaoPatrimonio") == "AUMENTOU").cast("double"))
    .withColumn("evolucaoPatrimonioDIMINUIU", (col("evolucaoPatrimonio") == "DIMINUIU").cast("double"))
    .withColumn("evolucaoPatrimonioMANTEVE", (col("evolucaoPatrimonio") == "MANTEVE").cast("double"))
)

dt_treino = (
    dt_treino
    .withColumn("perfilInvestidorSEM_INFORMACAO", (col("perfilInvestidor") == "SEM INFORMACAO").cast("double"))
    .withColumn("perfilInvestidorDIVERSIFICADO", (col("perfilInvestidor") == "DIVERSIFICADO").cast("double"))
    .withColumn("perfilInvestidorCONSERVADOR", (col("perfilInvestidor") == "CONSERVADOR").cast("double"))
    .withColumn("perfilInvestidorAGRESSIVO", (col("perfilInvestidor") == "AGRESSIVO").cast("double"))
)

dt_treino = dt_treino.drop("historicoAtrasoFinanciamentoVeicular","evolucaoPatrimonio","perfilInvestidor")

# VALIDACAO
colunas_string = [f.name for f in dt_treino.schema.fields if f.dataType.simpleString() == 'string']

for i in colunas_string:
    print("-------------",i,"-------------")
    dt_treino.groupby(i).count().show()
    print()

## Tratar coluna booleans

colunas_boolean = [f.name for f in dt_treino.schema.fields if f.dataType.simpleString() == 'boolean']

for coluna in colunas_boolean:
    dt_treino = dt_treino.withColumn(coluna, sf.col(coluna).cast("int"))

# VALIDACAO
colunas_boolean = [f.name for f in dt_treino.schema.fields if f.dataType.simpleString() == 'boolean']

for coluna in colunas_boolean:
    print("-------------",coluna,"-------------")
    dt_treino.groupby(coluna).count().show()
    print()

## Colunas Nulas

def countMissingValues(dataframe: DataFrame) -> DataFrame:
    missing_values = dataframe.select([
        sf.count(sf.when(sf.col(c).isNull(), c)).alias(c) for c in dataframe.columns
    ])

    return missing_values

countMissingValues(dt_treino).show(vertical=True, truncate=False)

dt_treino = dt_treino.fillna(0)

countMissingValues(dt_treino).show(vertical=True, truncate=False)

dt_treino.printSchema()

print("BAIXA RENDA")
(dt_treino
 .filter(sf.col('flagBaixaRenda')==1)
 .drop("flagBaixaRenda","flagMediaRenda","flagAltaRenda")
 .write.mode("overwrite")
 .parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean_baixaRenda.parquet")
)

print("MEDIA RENDA")
(dt_treino
 .filter(sf.col('flagMediaRenda')==1)
 .drop("flagBaixaRenda","flagMediaRenda","flagAltaRenda")
 .write.mode("overwrite")
 .parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean_mediaRenda.parquet")
)

print("ALTA RENDA")
(dt_treino
 .filter(sf.col('flagAltaRenda')==1)
 .drop("flagBaixaRenda","flagMediaRenda","flagAltaRenda")
 .write.mode("overwrite")
 .parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean_altaRenda.parquet")
)

import pandas as pd
pd.set_option('display.max_columns', None)

dt_treino.limit(5).toPandas()

dt_treino.write.mode("overwrite").parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean_com_cpf.parquet")

dt_treino.count()
