==> Perfil_Laranja/05_filter_(1).ipynb <==
# Importando dependencias

import pyspark.sql.functions as sf
from pyspark.sql.window import Window
from pyspark.sql.types import IntegerType
from feature_store import FeatureStore, Catalog
from pyspark.sql import DataFrame
import time
from pyspark.sql.functions import col, when, lit

# Pré-Tratamento dos Dados

df = spark.read.parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean.parquet")

total = df.count()

total

print(df.filter(sf.col("flagBaixaRenda") == 1).count())

print(df.filter(sf.col("flagMediaRenda") == 1).count())

print(df.filter(sf.col("flagAltaRenda") == 1).count())

from functools import reduce

cols = df.columns

# Cria a expressão que soma 1 para cada zero encontrado em uma linha
zero_count_expr = reduce(
    lambda a, b: a + b,
    [when(col(c) == 0, 1).otherwise(0) for c in cols]
)

df = df.withColumn("qtd_zeros", zero_count_expr)

df_group = (
    df.groupBy('qtd_zeros')
    .agg(sf.count('*').alias('count'))
    .withColumn('proportion', sf.round((sf.col('count') / sf.lit(total))*100, 5))
    .sort("qtd_zeros", ascending=False)
)

janela = Window.orderBy(sf.desc("qtd_zeros")).rowsBetween(Window.unboundedPreceding, Window.currentRow)

# Soma acumulada da coluna "x"
df_group = df_group.withColumn("proportion_acumulado", sf.sum("proportion").over(janela))

df_group.show()

df.filter(sf.col('qtd_zeros')<100).count()

(13674989/total)*100

(17543189/total)*100

baixa_renda = df.filter(sf.col("flagBaixaRenda") == 1)
media_renda = df.filter(sf.col("flagMediaRenda") == 1)
alta_renda = df.filter(sf.col("flagAltaRenda") == 1)

print(baixa_renda.filter(sf.col('qtd_zeros')<100).count())

print(media_renda.filter(sf.col('qtd_zeros')<100).count())

print(alta_renda.filter(sf.col('qtd_zeros')<100).count())

print("BAIXA RENDA")
(baixa_renda.filter(sf.col('qtd_zeros')<100)
 .drop("qtd_zeros")
 .write.mode("overwrite")
 .parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean_baixaRenda_reduzido.parquet")
)
print("MEDIA RENDA")
(media_renda.filter(sf.col('qtd_zeros')<100)
 .drop("qtd_zeros")
 .write.mode("overwrite")
 .parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean_mediaRenda_reduzido.parquet")
)
print("ALTA RENDA")
(alta_renda.filter(sf.col('qtd_zeros')<100)
 .drop("qtd_zeros")
 .write.mode("overwrite")
 .parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean_altaRenda_reduzido.parquet")
)

print("BAIXA RENDA")
(df
 .filter(sf.col('flagBaixaRenda')==1)
 .drop("flagBaixaRenda","flagMediaRenda","flagAltaRenda")
 .write.mode("overwrite")
 .parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean_baixaRenda.parquet")
)

print("MEDIA RENDA")
(df
 .filter(sf.col('flagMediaRenda')==1)
 .drop("flagBaixaRenda","flagMediaRenda","flagAltaRenda")
 .write.mode("overwrite")
 .parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean_mediaRenda.parquet")
)

print("ALTA RENDA")
(df
 .filter(sf.col('flagAltaRenda')==1)
 .drop("flagBaixaRenda","flagMediaRenda","flagAltaRenda")
 .write.mode("overwrite")
 .parquet("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean_altaRenda.parquet")
)
