==> Perfil_Laranja/06_Isolation_baixa.ipynb <==
# Importando dependencias

pip install gcsfs

import pyspark.sql.functions as sf
from pyspark.sql.window import Window
from pyspark.sql.types import IntegerType, StringType
from pyspark.sql import DataFrame

import gcsfs
import json
import pyarrow.parquet as pq

from functools import reduce
import subprocess
import os

from feature_store import FeatureStore, Catalog

import pandas as pd
import numpy as np

pd.set_option('display.max_columns', 150)

# Utils

import gcsfs
import pyarrow.parquet as pq
def read_parquet_to_pandas(path: str, projectGCS="analytics-k8s-dev-4742"):
    """function to read parquet file as
     as pandas dataframe - useful for
     training tasks"""
    fs = gcsfs.GCSFileSystem(project=projectGCS, requester_pays=True)
    files = ["gs://" + path for path in fs.glob(path + "/part-*")]
    df_pandas = pq.ParquetDataset(files, filesystem=fs).read().to_pandas()
    return df_pandas

BASE_GS = "gs://oculto/loss_prevention/perfil_laranja/data/"

df = read_parquet_to_pandas("gs://oculto/loss_prevention/perfil_laranja/data/perfil_laranja_clean_baixaRenda_reduzido.parquet")

df.shape

# Modelagem

from sklearn.ensemble import IsolationForest
from sklearn.metrics import mean_squared_error

pfs_sem_patrimonio_cond = (
    (df["qtdImoveis"] == 0) &
    (df["qtdVeiculosFinanciados"] == 0) &
    (df["quantidadeVeiculosPesados"] == 0) &
    (df["quantidadePropriedades"] == 0) &
    (df["quantidadeAeronaves"] == 0)
)

df = df[~pfs_sem_patrimonio_cond]

df.shape

df_sample = df.sample(frac=1, random_state=2025)

X = df_sample.copy()
X.shape

1==1

n_estimators = 100
contamination = 0.001
sample_size = 256

iso_forest = IsolationForest(n_estimators=n_estimators,
                            contamination=contamination,
                            max_samples=sample_size,
                            random_state=42)

%%time
iso_forest.fit(X)

df_sample['scoreAnomalia'] = iso_forest.decision_function(X)
df_sample['anomalia'] = iso_forest.predict(X)

df_sample['anomalia'].value_counts()

columns = X.columns

A importância é baseada na variação dos scores de anomalia ao embaralhar cada variável.

%%time
original_scores = iso_forest.decision_function(X)


feature_importance = {}

for feature in X.columns:
    print(feature)
    df_permuted = X.copy()
    df_permuted[feature] = np.random.permutation(df_permuted[feature])

    permuted_scores = iso_forest.decision_function(df_permuted)
    score_change = mean_squared_error(original_scores, permuted_scores)  # Impacto no score de anomalia

    feature_importance[feature] = score_change

max_importance = max(feature_importance.values())
for feature in feature_importance:
    feature_importance[feature] = (feature_importance[feature] / max_importance) * 100


df_importance = pd.DataFrame(
    list(feature_importance.items()), columns=["Feature", "Importance"]
).sort_values(by="Importance", ascending=False)

df_importance["Importance"] = df_importance["Importance"].round(2)

df_importance = pd.DataFrame(
    list(feature_importance.items()), columns=["Feature", "Importance"]
).sort_values(by="Importance", ascending=False)

df_importance.head(60)

importance_dict = df_importance.set_index("Feature")["Importance"].to_dict()

fs = gcsfs.GCSFileSystem(project="seu-projeto-gcp")

path = "gs://oculto/loss_prevention/perfil_laranja/data/df_importance_baixa_renda.json"

with fs.open(path, 'w') as f:
    json.dump(importance_dict, f)


spark.createDataFrame(df_importance).write.mode("overwrite").parquet(
    "gs://oculto/loss_prevention/perfil_laranja/data/df_importance_baixa_renda.parquet"
)
